[Box] xDAWN Trainer: Fix linux compilation (remove size_t)
Tracker Reference: [Jira CORE-428](https://jira.mensiatech.com/browse/CORE-428)

+++ plugins/processing/signal-processing/src/box-algorithms/ovpCBoxAlgorithmXDAWNTrainer.cpp
@@ -140,7 +140,7 @@ boolean CBoxAlgorithmXDAWNTrainer::process(void)
 
 		// Decodes input signals
 
-		for (size_t j = 0; j < 2; j++)
+		for (unsigned int j = 0; j < 2; j++)
 		{
 			n[j] = 0;
